# frontend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Book해도. REST API",
    "version": "2.0.0",
    "description": "쿠키 인증, 소유권 검증, 날짜별 optimistic concurrency. 전체 일정/메모/대안 변경에 expectedRevision 필수(누락 428, 충돌 409). POST 장소 추가는 원자적 append. X-Request-ID로 오류 추적."
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "동일 출처 API"
    }
  ],
  "paths": {
...
    "/weather": {
      "get": {
        "summary": "여행 날짜의 일별 날씨",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Weather"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 입력",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않거나 다른 사용자의 리소스",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한. Retry-After 후 재시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "minimum": 41,
              "maximum": 46.1
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "minimum": 137,
              "maximum": 147
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "description": "Open-Meteo 제공. 일본 시간 오늘부터 10일 이내만 조회. 예보 범위 밖 또는 외부 서비스 오류 시 available=false와 안내문 반환.",
        "tags": [
          "날씨"
        ],
        "operationId": "get__weather"
      }
    },
...
    "/trips/{id}/days/{date}/weather-alternatives": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "date",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "date"
          }
        }
      ],
      "get": {
        "summary": "날씨 기반 실내 대안 후보와 선택 후보의 경로 미리보기",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherAlternativeResult"
                }
              }
            }
          },
          "400": {
            "description": "입력 형식 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "본인 여행·날짜·대상 장소 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "추천 후보 변경",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한. Retry-After 후 재시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "targetId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "previewId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "description": "소유권 검증 후 대상 장소 좌표의 Open-Meteo 예보를 조회한다. 비/눈/이슬비/뇌우 또는 일 강수량 5mm, 적설 1cm, 강수확률 70%, 풍속 40km/h 이상 중 하나이면 실내 대안 탐색. 공식 기상특보 기준이 아님. 실내 분류 및 직선 20km 이내 후보를 기존 일정에서 제외하고 최대 3곳 제시. previewId는 현재 추천 후보만 허용. 앞뒤 인접 구간의 실제 경로 조회 실패 시 비교 시간 null. 저장하지 않음.",
        "tags": [
          "추천"
        ],
        "operationId": "get__trips_id_days_date_weather_alternatives"
      },
      "patch": {
        "summary": "사용자 확정 후 야외 장소를 주변 실내 장소로 교체",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "saved": {
                      "type": "boolean"
                    },
                    "revision": {
                      "type": "integer",
                      "minimum": 0,
                      "description": "조회 시 받은 날짜 버전. 일정·메모가 바뀌면 증가"
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 입력·실내/거리 조건 불충족",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "본인 여행·날짜 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "조회 이후 일정·메모가 변경됨. 최신 내용 확인 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "428": {
            "description": "expectedRevision 누락",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한. Retry-After 후 재시도",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "저장 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "targetId",
                  "replacementId",
                  "placeIds",
                  "expectedRevision"
                ],
                "properties": {
                  "targetId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "replacementId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "placeIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "maxItems": 30,
                    "uniqueItems": true,
                    "description": "미리보기 시점 기존 일정 전체의 ID 순서 (expectedPlaceIds)"
                  },
                  "expectedRevision": {
                    "type": "integer",
                    "minimum": 0,
                    "description": "조회 시 받은 날짜 버전. 일정·메모가 바뀌면 증가"
                  }
                }
              }
            }
          }
        },
        "description": "날짜 행을 잠그고 기존 순서와 비교해 동시 수정 409를 반환한다. 실내 분류·20km·중복을 재검증하며 예보는 저장 시 다시 호출하지 않는다. 기존 순번과 다른 항목의 메모를 유지하고 교체 대상 메모만 비운다. 오류 시 rollback. 추가 테이블 없이 날짜의 revision을 사용한다. expectedRevision을 잠긴 날짜의 현재 revision과 비교하며 성공하면 revision을 증가시킨다.",
        "tags": [
          "추천"
        ],
        "operationId": "patch__trips_id_days_date_weather_alternatives"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kita_session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "currentRevision": {
            "type": "integer",
            "minimum": 0,
            "description": "조회 시 받은 날짜 버전. 일정·메모가 바뀌면 증가"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Place": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "regionId": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "ATTRACTION",
              "RESTAURANT",
              "LODGING"
            ]
          },
          "name": {
            "type": "string"
          },
          "nameJa": {
            "type": "string"
          },
          "nameKo": {
            "type": [
              "string",
              "null"
            ]
          },
          "nameEn": {
            "type": [
              "string",
              "null"
            ]
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "address": {
            "type": [
              "string",
              "null"
            ]
          },
          "municipality": {
            "type": [
              "string",
              "null"
            ]
          },
          "website": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "openingHours": {
            "type": [
              "string",
              "null"
            ]
          },
          "tags": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "sources": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "source": {
                  "type": "string"
                },
                "url": {
                  "type": "string"
                },
                "license": {
                  "type": "string"
                }
              }
            }
          },
          "operatingPeriod": {
            "type": "null"
          },
          "seasonNotice": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "trendBadge": {
            "type": [
              "string",
              "null"
            ]
          },
          "reviewSummary": {
            "type": "object",
            "properties": {
              "rating": {
                "type": "number"
              },
              "reviewCount": {
                "type": "integer"
              },
              "japaneseCount": {
                "type": "integer"
              },
              "sampleCount": {
                "type": "integer"
              },
              "score": {
                "type": "number"
              }
            }
          },
          "recommended": {
            "type": "boolean"
          },
          "recommendationReasons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "estimatedCost": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0,
            "maximum": 100000000,
            "description": "사용자가 입력한 장소 예상 총액 (JPY). null은 미입력."
          },
          "startMinute": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0,
            "maximum": 1439
          },
          "endMinute": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1,
            "maximum": 1440
          }
        }
      },
...
      "Segment": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "format": "uuid"
          },
          "to": {
            "type": "string",
            "format": "uuid"
          },
          "distanceMeters": {
            "type": "number"
          },
          "durationSeconds": {
            "type": [
              "number",
              "null"
            ]
          },
          "coordinates": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          },
          "estimatedCost": {
            "type": "null"
          },
          "costNotice": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "transitFare": {
            "type": [
              "string",
              "object",
              "null"
            ]
          },
          "tolls": {
            "type": [
              "object",
              "null"
            ]
          },
          "costs": {
            "type": [
              "object",
              "null"
            ]
          },
          "polyline": {
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "valhalla",
              "google",
              "straight-line"
            ]
          },
          "notice": {
            "type": "string"
          }
        }
      },
      "Weather": {
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "high": {
            "type": "number"
          },
          "low": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "sourceUrl": {
            "type": "string",
            "format": "uri"
          },
          "notice": {
            "type": "string"
          },
          "precipitationMm": {
            "type": [
              "number",
              "null"
            ]
          },
          "snowfallCm": {
            "type": [
              "number",
              "null"
            ]
          },
          "precipitationProbability": {
            "type": [
              "number",
              "null"
            ]
          },
          "windSpeedKmh": {
            "type": [
              "number",
              "null"
            ],
            "description": "일 최고 10m 풍속 km/h"
          }
        }
      },
      "WeatherAlternative": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Place"
          },
          {
            "type": "object",
            "properties": {
              "indoorEvidence": {
                "type": "string"
              },
              "distanceMeters": {
                "type": "number",
                "description": "기존 장소에서의 직선거리 m"
              },
              "detourMeters": {
                "type": "number",
                "description": "인접 구간 직선거리 변화 m"
              },
              "operationNotice": {
                "type": "string"
              }
            }
          }
        ]
      },
      "WeatherAlternativeResult": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "READY",
              "NO_FORECAST",
              "FAIR_WEATHER",
              "NOT_OUTDOOR",
              "NO_CANDIDATES"
            ]
          },
          "weather": {
            "$ref": "#/components/schemas/Weather"
          },
          "target": {
            "$ref": "#/components/schemas/Place"
          },
          "expectedPlaceIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeatherAlternative"
            }
          },
          "notice": {
            "type": "string"
          },
          "preview": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "properties": {
                  "candidate": {
                    "$ref": "#/components/schemas/WeatherAlternative"
                  },
                  "before": {
                    "type": "object",
                    "properties": {
                      "segments": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Segment"
                        }
                      },
                      "complete": {
                        "type": "boolean"
                      },
                      "distanceMeters": {
                        "type": [
                          "number",
                          "null"
                        ]
                      },
                      "durationSeconds": {
                        "type": [
                          "number",
                          "null"
                        ]
                      }
                    }
                  },
                  "after": {
                    "type": "object",
                    "properties": {
                      "segments": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Segment"
                        }
                      },
                      "complete": {
                        "type": "boolean"
                      },
                      "distanceMeters": {
                        "type": [
                          "number",
                          "null"
                        ]
                      },
                      "durationSeconds": {
                        "type": [
                          "number",
                          "null"
                        ]
                      }
                    }
                  },
                  "durationDeltaSeconds": {
                    "type": [
                      "number",
                      "null"
                    ]
                  }
                }
              }
            ]
          },
          "tripId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "expectedRevision": {
            "type": "integer",
            "minimum": 0,
            "description": "조회 시 받은 날짜 버전. 일정·메모가 바뀌면 증가"
          }
        }
      },
...
    }
  }
}